refactor inclusion report types
