+++ ui/apps/vue-mri-ui-lib/src/query-filter/components/InclusionReport/computeAttritionStats.ts
@@ -1,4 +1,4 @@
-import type { InclusionReportResponse } from '@/query-filter/types/QueryFilterTypes'
+import type { InclusionReportResponse } from '@/query-filter/types/InclusionReportTypes'
 
 type AttritionStat = {
   id: number
