+++ ui/apps/vue-mri-ui-lib/src/query-filter/components/InclusionReport/computeTreemapStats.ts
@@ -1,4 +1,4 @@
-import { InclusionReportResponse } from '@/query-filter/types/QueryFilterTypes'
+import { InclusionReportResponse } from '@/query-filter/types/InclusionReportTypes'
 
 type TooltipData = {
   count: string
